feat: add macOS and Windows installers for sshforward
- Introduced a macOS installer using productbuild with a distribution XML and welcome HTML.
- Created an Inno Setup script for a Windows installer that supports multiple architectures.
- Added tests to ensure installer configurations are correct and files exist.
- Updated package identifiers and repository references from pcpl2 to pcpl2lab.
- Enhanced source detection for executables to recognize installations from the Windows installer and macOS package.
- Included a Windows icon resource and updated versioning in the winres configuration.

+++ internal/tunnel/process_windows.go
@@ -16,8 +16,11 @@ import (
 const stillActive = 259
 
 func IsProcessAlive(pid int) bool {
-	if pid <= 0 || pid > math.MaxUint32 {
-		return false // state files are user-editable; reject PIDs outside DWORD range
+	// State files are user-editable, so reject PIDs outside the DWORD range that
+	// OpenProcess takes. The uint64 conversion is what makes this compile on
+	// 386, where int is 32 bits and cannot be compared to MaxUint32 directly.
+	if pid <= 0 || uint64(pid) > math.MaxUint32 {
+		return false
 	}
 	h, err := windows.OpenProcess(windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
 	if err != nil {
@@ -36,7 +39,7 @@ func IsProcessAlive(pid int) bool {
 }
 
 func IsSSHProcess(pid int) bool {
-	if pid <= 0 || pid > math.MaxUint32 {
+	if pid <= 0 || uint64(pid) > math.MaxUint32 {
 		return false
 	}
 	h, err := windows.OpenProcess(windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
